Generic command bus
The provided command bus was refactored in order to make it
cross-framework.

It uses a new Container interface to resolve classes out of a
framework's container. A Laravel implementation for the container is
provided.

Because of these changes, the command bus now lives in StudioIgnis\Cmd
namespace.

This shouldn't be a breaking change.